+++ include/core.hpp
@@ -2,7 +2,6 @@
 
 #include <chrono>
 #include <filesystem>
-#include <format>
 #include <functional>
 #include <iostream>
 #include <unordered_map>
